fleshing out routes and logic

+++ Unit-02/01-blueprints/project/users/views.py
@@ -14,7 +14,7 @@ def index():
 	if request.method == 'POST':
 		user_form = UserForm(request.form)
 		if user_form.validate():
-			new_user = User(request.form['first_name'], request.form['last_name'])
+			new_user = User(user_form.data.first_name, user_form.data.last_name)
 			db.session.add(new_user)
 			db.session.commit()
 			return redirect(url_for('users.index'))
@@ -31,17 +31,19 @@ def show(id):
 	if request.method == b'PATCH':
 		user_form = UserForm(request.form)
 		if user_form.validate():
-			user.first_name = request.form['first_name']
-			user.last_name = request.form['last_name']
+			user.first_name = user_form.data.first_name
+			user.last_name = user_form.data.last_name
 			db.session.add(user)
 			db.session.commit()
 			return redirect(url_for('user.show'))
 		else:
 			return render_template('edit.html', form=user_form, user=user)
 	elif request.method == b'DELETE':
-		db.session.delete(user)
-		db.session.commit()
-		return redirect(url_for('user.index'))			
+		delete_form = DeleteForm(request.form)
+		if delete_form.validate():
+			db.session.delete(user)
+			db.session.commit()
+			return redirect(url_for('user.index'))			
 	return render_template('show.html', user=user)
 
 @users_blueprint.route('/<int:id>/edit')
